refactor: move platforms, icons, splash to constants and config
refactor: move image formats to constants

refactor: move display helper to utils

refactor: move fs related methods to core/fs-manager

refactor: move methods to core modules

refactor: move some methods to utils module
