+++ src/config/icons/android.js
