+++ src/config/icons/blackberry10.js
