+++ src/config/icons/pwa.js
