+++ src/config/icons/windows.js
