+++ src/config/splash/android.js
