+++ src/config/splash/ios.js
