+++ src/config/splash/windows.js
