+++ src/constants/image-formats.js
@@ -1,0 +1,13 @@
+exports.IMAGE_FORMATS = [
+  'svg',
+  'webp',
+  'png',
+  'tif',
+  'tiff',
+  'dzi',
+  'szi',
+  'v',
+  'vips',
+  'jpg',
+  'jpeg',
+];
