+++ src/constants/platforms.js
@@ -1,0 +1,19 @@
+// app main constiables and constants
+const PLATFORMS = {
+  android: {
+    definitions: ['../config/icons/android', '../config/splash/android'],
+  },
+  ios: {
+    definitions: ['../config/icons/ios', '../config/splash/ios'],
+  },
+  windows: {
+    definitions: ['../config/icons/windows', '../config/splash/windows'],
+  },
+  blackberry10: {
+    definitions: ['../config/icons/blackberry10'],
+  },
+  pwa: {
+    definitions: ['../config/icons/pwa.js'],
+  },
+};
+module.exports = PLATFORMS;
