+++ src/utils/error-handlers.js
@@ -1,0 +1,8 @@
+function catchErrors(err) {
+  if (err) {
+    console.log('Error: ', err.message);
+    process.exit(1);
+  }
+}
+
+exports.catchErrors = catchErrors;
